delta_lake_handler/requirements.txt
pyspark==3.2.1
delta-spark==1.2.1
pydantic==1.10.9